refactor: move methods to related functions files with tests also delete unused array.go and array_test.go
