+++ array/findLastIndex.go
@@ -10,3 +10,8 @@ func FindLastIndex[T comparable](slice []T, fn Predicate[T]) int {
 
 	return -1
 }
+
+// The FindLastIndex() method returns the index of the last element that passes the provided predicate function.
+func (a Array[T]) FindLastIndex(fn Predicate[T]) int {
+	return FindLastIndex([]T(a), fn)
+}
