+++ array/flat.go
@@ -61,3 +61,16 @@ func Flat[T any](slice []any, depths ...int) ([]T, error) {
 	}
 	return result, nil
 }
+
+// The Flat() method flattens nested arrays up to the specified depth.
+// The Flat() method returns a new array with the sub-array elements concatenated into it. The depth parameter specifies how deep a nested array structure should be flattened. The default is 1.
+// Until Generic Methods are implemented in Go, the Flat() method can only return Array[any] since it needs to handle slices of any type. Once Generic Methods are available, we can update the Flat() method to return Array[T] and handle type assertions accordingly.
+func (a Array[T]) Flat(depths ...int) (Array[any], error) {
+	// Convert Array[T] to []any for flattening
+	anySlice := make([]any, len(a))
+	for i, v := range a {
+		anySlice[i] = v
+	}
+	result, err := Flat[any](anySlice, depths...)
+	return Array[any](result), err
+}
