+++ array/forEach.go
@@ -5,3 +5,8 @@ func ForEach[T comparable](slice []T, fn ForEachFunc[T]) {
 		fn(value, index, slice)
 	}
 }
+
+// The ForEach() method executes a provided function once for each array element.
+func (a Array[T]) ForEach(fn ForEachFunc[T]) {
+	ForEach([]T(a), fn)
+}
