+++ array/push.go
@@ -4,3 +4,8 @@ package array
 func Push[T comparable](slice []T, elements ...T) []T {
 	return append(slice, elements...)
 }
+
+// The Push() method adds one or more elements to the end of the array and returns the new array.
+func (a Array[T]) Push(elements ...T) Array[T] {
+	return Array[T](Push([]T(a), elements...))
+}
