+++ array/reduceRight.go
@@ -64,3 +64,13 @@ func ReduceRightStrict[T comparable](slice []T, fn ReduceStrictFunc[T], initialV
 
 	return accumulator, nil
 }
+
+// The ReduceRight() method executes a reducer function for each array element (from right to left) and returns a single value.
+func (a Array[T]) ReduceRight(fn ReduceFunc[T], initialValue any) (any, error) {
+	return ReduceRight([]T(a), fn, initialValue)
+}
+
+// The ReduceRightStrict() method executes a reducer function for each array element (from right to left) and returns a value of the same type.
+func (a Array[T]) ReduceRightStrict(fn ReduceStrictFunc[T], initialValue *T) (T, error) {
+	return ReduceRightStrict([]T(a), fn, initialValue)
+}
