+++ array/with.go
@@ -16,3 +16,9 @@ func With[T comparable](slice []T, index int, value T) ([]T, error) {
 
 	return newSlice, nil
 }
+
+// The With() method returns a new array with the element at the given index replaced with the given value.
+func (a Array[T]) With(index int, value T) (Array[T], error) {
+	result, err := With([]T(a), index, value)
+	return Array[T](result), err
+}
